[fe] refact: NoticeBoardWrite에서 Component화 진행
